Update: Worlde copy
Heheheh... there was a minor bug which caused the same word appearing more than once. I fixed it, and I do warn you, do NOT input anything over the amount of words in the possible answer. The game will get stuck in an infinite loop, and throw an exception because it is a recursive function.